pathnames: widen slab_off and len to size_t in findex builder
The second-pass namelist->findex builder tracks the slab write offset
with `unsigned int slab_off` and sizes each string copy with
`unsigned int len`, while the measuring accumulator total_str_bytes is
already size_t. Once the pathname pool exceeds 4 GiB, `slab_off += len`
wraps at UINT_MAX and the following memcpy scribbles earlier strings.
Widen both to size_t so the write cursor matches the accumulator.

+++ pathnames.c
@@ -466,7 +466,7 @@ static const char ** list_to_index(struct namelist *namelist)
 	unsigned int i = 0;
 	size_t total_str_bytes = 0;
 	char *slab;
-	unsigned int slab_off = 0;
+	size_t slab_off = 0;
 
 	/* First pass: measure total string storage needed. */
 	list_for_each(node, &namelist->list) {
@@ -489,7 +489,7 @@ static const char ** list_to_index(struct namelist *namelist)
 	/* Second pass: copy strings into the slab and build the index. */
 	list_for_each_safe(node, tmp, &namelist->list) {
 		nl = (struct namelist *) node;
-		unsigned int len = strlen(nl->name) + 1;
+		size_t len = strlen(nl->name) + 1;
 
 		memcpy(slab + slab_off, nl->name, len);
 		findex[i++] = slab + slab_off;
